+++ tests/test_provider.py
@@ -807,7 +807,7 @@ class TestNormalizeValidateDenormalize:
                     "selector": None,
                     "operator": "IPMatch",
                     "negate_condition": False,
-                    "match_value": ["203.0.113.0/24"],
+                    "match_value": ["104.16.0.0/12"],
                     "transforms": [],
                 }
             ],
@@ -875,7 +875,7 @@ class TestConcurrentWorkers:
                                 "selector": None,
                                 "operator": "IPMatch",
                                 "negate_condition": False,
-                                "match_value": ["203.0.113.0/24"],
+                                "match_value": ["104.16.0.0/12"],
                                 "transforms": [],
                             }
                         ],